requirements.txt
Django==5.1.4
mssql-django==1.5
pyodbc==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
gunicorn==23.0.0
Pillow==11.1.0
openpyxl==3.1.5
reportlab==4.2.5
django-storages==1.14.4
boto3==1.35.0
python-decouple==3.8
bcrypt==4.2.1
djangorestframework-camel-case==1.4.2
drf-spectacular==0.28.0
whitenoise==6.8.2
django-extensions==3.2.3
